Update ball collision logic to fix a certain class of bugs
There are two cases that could cause hangs. The first is where the ball hits the top or bottom of the paddle while the paddle is moving vertically against it with a greater magnitude. This means when the ball's y-velocity is reflected the paddle is still moving faster than it. This is resolved by adding the paddle's y-velocity to the ball to ensure the ball is moving away from it faster than the ball is moving.

The second is when the ball gets pinched between a wall and a paddle. This would cause it to continually try to reflect between the two and become endlessly stuck. Resolved this by just setting the y velocity to zero so it could slide between the other two objects.